+++ resource.h
@@ -1,5 +1,7 @@
 #pragma once
 #define IDI_APPICON 101
 #define IDB_ARTORIAS 102
+#define IDB_BG_DARK  103
+#define IDB_BG_LIGHT 104
 #define IDR_CHANGELOG_LAUNCHER  200
 #define IDR_CHANGELOG_MCTDELINK 201